+++ lib/any_borders.dart
@@ -2,6 +2,7 @@ library any_borders;
 
 export 'any_contour.dart';
 export 'any_box_decoration.dart';
+export 'any_decoration_cache.dart';
 export 'any_decoration_tween.dart';
 export 'any_fill.dart';
 export 'any_shadow.dart';